requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
black==21.5b1
bleach==3.3.0
cffi==1.14.5
click==8.0.1
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
future==0.18.2
husl==4.0.3
importlib-metadata==4.3.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.4
packaging==20.9
pandas==1.1.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
pycrypto==2.6.1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.0
secretpy==0.9.0
Send2Trash==1.5.0
six==1.16.0
sklearn==0.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.0
torchmetrics==0.2.0
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
